requirement.txt
absl-py==1.1.0
addict==2.4.0
anyio==3.6.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
attrs==21.4.0
Babel==2.10.3
backcall==0.2.0
basicsr==1.3.4.9
beautifulsoup4==4.11.1
bleach==5.0.1
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cycler==0.11.0
debugpy==1.6.2
decorator==5.1.1
defusedxml==0.7.1
einops==0.6.1
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.15.3
filelock==3.12.2
fonttools==4.34.4
fsspec==2023.6.0
future==0.18.3
google-auth==2.22.0
google-auth-oauthlib==1.0.0
grpcio==1.57.0
h5py==3.9.0
huggingface-hub==0.16.4
idna==3.3
imageio==2.31.1
importlib-metadata==6.8.0
importlib-resources==5.8.0
ipykernel==6.15.1
ipython==8.4.0
ipython-genutils==0.2.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
json5==0.9.8
jsonschema==4.7.2
jupyter-client==7.3.4
jupyter-core==4.11.1
jupyter-server==1.18.1
jupyterlab==3.4.3
jupyterlab-language-pack-zh-CN==3.4.post1
jupyterlab-pygments==0.2.2
jupyterlab-server==2.15.0
keras==2.9.0
kiwisolver==1.4.3
lazy_loader==0.3
lmdb==1.4.1
Markdown==3.3.7
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mistune==0.8.4
nbclassic==0.4.3
nbclient==0.6.6
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==3.1
notebook-shim==0.1.0
numpy==1.22.3
nvidia-ml-py==11.495.46
nvitop==0.6.2
oauthlib==3.2.0
opencv-contrib-python==4.5.5.64
opencv-python==4.8.0.76
packaging==21.3
pandas==1.4.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pip==22.0.4
platformdirs==3.10.0
prometheus-client==0.14.1
prompt-toolkit==3.0.30
protobuf==4.24.1
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
py3nvml==0.2.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.12.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
pytorch-msssim==1.0.0
pytz==2022.1
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq==23.2.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.8
safetensors==0.3.3
scikit-image==0.21.0
scikit-learn==1.1.1
scipy==1.8.1
Send2Trash==1.8.0
setuptools==62.1.0
six==1.16.0
sklearn==0.0
sniffio==1.2.0
soupsieve==2.3.2.post1
stack-data==0.3.0
tb-nightly==2.14.0a20230522
tensorboard==2.9.1
tensorboard-data-server==0.7.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6.2.2
termcolor==1.1.0
terminado==0.15.0
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tifffile==2023.7.10
timm==0.9.5
tinycss2==1.1.1
tomli==2.0.1
torch==1.12.0+cu113
torchaudio==0.12.0+cu113
torchvision==0.13.0+cu113
tornado==6.2
tqdm==4.66.1
traitlets==5.3.0
typing_extensions==4.3.0
urllib3==1.26.10
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.3
Werkzeug==2.1.2
wheel==0.37.1
xgboost==1.6.1
xmltodict==0.12.0
yapf==0.40.1
zipp==3.8.1
